+++ storage/lib/src/main/java/org/calyxos/backup/storage/prune/Pruner.kt
@@ -44,6 +44,7 @@ internal class Pruner(
         val duration = measure {
             val storedSnapshots = backend.getCurrentBackupSnapshots(androidId)
             val toDelete = retentionManager.getSnapshotsToDelete(storedSnapshots)
+            Log.i(TAG, "Found ${storedSnapshots.size} snapshots. Deleting ${toDelete.size}...")
             backupObserver?.onPruneStart(toDelete.map { it.timestamp })
             for (snapshot in toDelete) {
                 try {
@@ -51,6 +52,12 @@ internal class Pruner(
                 } catch (e: Exception) {
                     Log.e(TAG, "Error pruning $snapshot", e)
                     backupObserver?.onPruneError(snapshot.timestamp, e)
+                    if (e.message?.contains("BAD_DECRYPT") == true) try {
+                        Log.i(TAG, "Removing $snapshot because can't decrypt it")
+                        backend.remove(snapshot.snapshotHandle)
+                    } catch (ex: Exception) {
+                        Log.e(TAG, "Error deleting $snapshot", ex)
+                    }
                 }
             }
         }
@@ -80,7 +87,9 @@ internal class Pruner(
             size += it.size
             it.id
         }
-        backupObserver?.onPruneSnapshot(storedSnapshot.timestamp, chunkIdsToDelete.size, size)
+        val timestamp = storedSnapshot.timestamp
+        backupObserver?.onPruneSnapshot(timestamp, chunkIdsToDelete.size, size)
+        Log.d(TAG, "Removed $timestamp, now deleting ${chunkIdsToDelete.size} chunks...")
         chunkIdsToDelete.forEach { chunkId ->
             backend.remove(FileBackupFileType.Blob(androidId, chunkId))
         }